refactor: move FileRules code to StrictRules\FileRules
StrictRules are now the default.
